refactor: separate types file
