+++ src/shared/ui/nav-item/NavItem.tsx
@@ -1,10 +1,21 @@
 "use client"
 
-import { MenuItemDataType, MenuItemType } from '@/shared/types/types';
 import styles from './NavItem.module.scss';
 import Image, { StaticImageData } from 'next/image';
 import Link from 'next/link';
 import { useState } from 'react';
+
+export type MenuItemDataType = {
+  id: string,
+  text: string,
+  img: StaticImageData,
+  imgWhite: StaticImageData
+};
+
+export type MenuItemType = {
+  id: string,
+  data: MenuItemDataType
+};
 
 export function NavItem({ data }: MenuItemType) {
   const [isCurElHover, setIsCurElHover] = useState(false);
