+++ src/shared/ui/tag/Tag.tsx
@@ -1,8 +1,11 @@
-import { IdValue } from '@/shared/types/types';
 import styles from './Tag.module.scss';
 
+export type IdValue = {
+  id: string,
+  value: string
+}
+
 export function Tag({ props }: { props: IdValue }) {
-  console.log('value: ', props.value);
   return (
     <div className={styles.tag}>{`#${props.value}`}</div>
   )
